requirements.txt
asgiref==3.3.0
Django==3.1.2
django-createsuperuser==0.0.0
django-filter==2.4.0
django-versatileimagefield==2.0
djangorestframework==3.12.1
env==0.1.0
Markdown==3.3.2
MarkupSafe==1.1.1
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.1
pycryptodome==3.9.8
pytesseract==0.3.6
python-dotenv==0.14.0
python-magic==0.4.18
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
gunicorn